reorg metadata collector code
